requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-environ==0.9.0
django-filter==21.1
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==3.0.9
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.6
whitenoise==5.3.0
